LayoutTests/imported/w3c/web-platform-tests/css/selectors/focus-visible-002-expected.txt

This test checks that :focus-visible always matches on <input> elements which take text input, regardless of focus mechanism.
If the user-agent does not claim to support the :focus-visible pseudo-class then SKIP this test.
Click each form element below to focus it.
If the element has a red background, then the test result is FAILURE. If the element has a green outline, then the test result is SUCCESS.

















FAIL Mouse focus on elements which would show a virtual keyboard should match :focus-visible assert_equals: expected "rgb(0, 100, 0)" but got "rgba(46, 52, 54, 0.59)"
